no va ni el int ni el float

+++ funciones.py
@@ -42,7 +42,7 @@ def funcion2():
             Ordinario2 = i[6]
             Practicas = i[7]
             OrdinarioPracticas = i[8]
-            notafinal = int(Parcial1) * 0.3 + int(Parcial2) * 0.3 + int(Ordinario1) * 0.3 + int(Ordinario2) * 0.3 + int(Practicas) * 0.4 + int(OrdinarioPracticas) * 0.4
+            notafinal = float(Parcial1) * 0.3 + float(Parcial2) * 0.3 + float(Ordinario1) * 0.3 + float(Ordinario2) * 0.3 + float(Practicas) * 0.4 + float(OrdinarioPracticas) * 0.4
         print(f"Nombre y Apellidos: '{Nombre} {Apellidos} con una nota final de {notafinal}")
         file.close()
 # float(i["Parcial1"])
